Add classroom and coordinate data
data.xlsx contains 625 14x7 sheets, with the name of each sheet corresponding to the name of the classroom. The tables have the columns "Hour", "M", "T", "W", "R", "F". the days of the week contain a string with either "true" or "false" to represent whether a class is currently occurring.

coordinates.xlsx contains 3 columns. building, lat, and long.